+++ tessera-ui-basic-components/src/scrollable.rs
@@ -576,7 +576,7 @@ fn scrollable_inner(
     }
 
     // Handle scroll input and position updates
-    state_handler(Box::new(move |input| {
+    input_handler(Box::new(move |input| {
         let size = input.computed_data;
         let cursor_pos_option = input.cursor_position_rel;
         let is_cursor_in_component = cursor_pos_option